rest-api/requirements.txt
Flask
flask-socketio
flask-jwt-extended
flask-sqlalchemy
psycopg2-binary
gunicorn[eventlet]
celery[redis]